Hoarder/requirements.txt
fastapi[all]
python-dotenv
mysql-connector-python
pika
pydantic==1.10.2
